+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.5] - 2024-11-20
+
+**Fix**:
+- Fixed the issues of *ObservableDictionary* when subscribing/unsubscribing to actions while removing/adding elements
+- Fixed the issues of *ObservableList* when subscribing/unsubscribing to actions while removing/adding elements
+
 ## [0.6.4] - 2024-11-13
 
 **Fix**: